+++ app/schema/academicprofiles_schema.py
@@ -14,7 +14,7 @@ class PostAcademicProfiles(BaseModel):
     type_institution: str | None = Field(description='Type_institution Documentar', max_length=45)
     teaching_modality: str | None = Field(description='Teaching_modality Documentar', max_length=45)
     other_courses: str | None = Field(description='Other_courses Documentar', max_length=255)
-    academic_coefficient: int | None = Field(description='Coeficiente Acadêmico, favor transformar em inteiro de 0 a 100')
+    academic_coefficient: float | None = Field(description='Coeficiente Acadêmico, favor transformar em inteiro de 0 a 100')
